requirements.txt
matplotlib==3.7.1
MIDIUtil==1.2.1
mido==1.2.10
numpy==1.24.3
openai==1.86.0
streamlit==1.28.2
python-dotenv>=1.0.0
pydantic>=2.6
